requirements.txt
playwright>=1.40.0
PyYAML>=6.0
requests>=2.28.0
APScheduler>=3.10.0
